Make event wrapping calculation more robust
The previous CSS-only solution didn't take into account multi-day
events. Attempting to fix this by making the CSS-only solution depend on
the style="margin-top: 25px;" attributes (that Full Calendar uses to
push down events to make room for multi-day events) leads to infinite
layout loops! So instead I've switched to a JavaScript solution that
counts how many other events overlap with each event (on the day that
the most events overlap with it), and sets a class name on the event
accordingly, giving a stable layout that accounts for multi-day events.

This still isn't quite perfect:
- On months when there are many overlapping events, the daygrid can
  require vertical scrolling. But that doesn't seem fixable without
  deep surgery in the Full Calendar layout code, so it'll have to do.
- Sometimes Full Calendar doesn't render all day events at the top of a
  day, but instead partway down a day, because on a previous day they
  overlapped with some other all day event. As a result the
  `classNamesForEvent` calculation occasionally undercounts the
  effective number of events taking up space on a day, and hence allows
  events to wrap onto more lines than they should.